requirements.txt
Flask==3.0.3
flask-cors==4.0.1
requests==2.31.0

flask
flask-cors
mysql-connector-python